{{backport int_master}} Do not memory mmap memory during hack replacement
KLEE uses a hack to fast-clear memory objects by replacing the
memory manager with a new one.
For deterministic memory, this executes mmap a second time.
Previously freed mmaped memory might not be returned therefore,
there is not enough memory available anymore
which leads to the next mmap fail.

As those mmap typically acquire a lot of memory (3GB) we avoid this
a second time.