Reduce coverage suite using work estimation
To reduce the risk that two test-cases cover the same path
we check if some symbolic variables have been involved in
recent control-flow decision.
If so, KLEE will have spent some time in the solver chain.
Using this fact, we will dump a state if it covers a new basic block
and has spent some time beforehand in the solver chain.
After we dumped a state, we reset its counter.

We are currently abusing the time counter.
We could also add an additional variable that is incremented as soon
as we reach the solver chain. This might be more correct as super-fast
solving chain access might not show as timer changes.